+++ src/main/java/sopt/makers/authentication/adapter/in/web/util/CookieUtil.java
@@ -3,7 +3,7 @@ package sopt.makers.authentication.adapter.in.web.util;
 import static sopt.makers.authentication.adapter.out.jwt.JwtConstant.REFRESH_TOKEN_HEADER;
 import static sopt.makers.authentication.common.constant.SystemConstant.PATTERN_ROOT_PATH;
 
-import sopt.makers.authentication.support.value.SecurityProperty;
+import sopt.makers.authentication.config.SecurityProperty;
 
 import java.time.Duration;
