+++ src/main/java/sopt/makers/authentication/application/service/user/GetSocialAccountPlatformService.java
@@ -2,10 +2,10 @@ package sopt.makers.authentication.application.service.user;
 
 import sopt.makers.authentication.application.port.in.auth.GetSocialAccountUsecase;
 import sopt.makers.authentication.application.port.out.user.UserRepository;
+import sopt.makers.authentication.application.validator.auth.PhoneVerificationValidator;
 import sopt.makers.authentication.domain.auth.AuthPlatform;
 import sopt.makers.authentication.domain.auth.PhoneVerificationType;
 import sopt.makers.authentication.domain.user.User;
-import sopt.makers.authentication.support.validator.PhoneVerificationValidator;
 
 import org.springframework.stereotype.Service;
