+++ src/main/java/sopt/makers/authentication/application/service/user/UpdateUserProfileService.java
@@ -5,6 +5,7 @@ import static sopt.makers.authentication.domain.user.exception.UserFailure.NOT_F
 import sopt.makers.authentication.application.port.in.user.UpdateUserProfileUsecase;
 import sopt.makers.authentication.application.port.out.user.UserActivityHistoryRepository;
 import sopt.makers.authentication.application.port.out.user.UserRepository;
+import sopt.makers.authentication.application.validator.auth.PhoneVerificationValidator;
 import sopt.makers.authentication.domain.auth.PhoneVerificationType;
 import sopt.makers.authentication.domain.user.Activity;
 import sopt.makers.authentication.domain.user.ActivityList;
@@ -12,7 +13,6 @@ import sopt.makers.authentication.domain.user.Profile;
 import sopt.makers.authentication.domain.user.Team;
 import sopt.makers.authentication.domain.user.User;
 import sopt.makers.authentication.domain.user.exception.UserException;
-import sopt.makers.authentication.support.validator.PhoneVerificationValidator;
 
 import java.util.List;
 import java.util.Map;
